server/pyproject.toml
[project]
name = "ai-interview-coach-server"
version = "0.1.0"
description = "AI Interview Coach Server"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.9.3",
    "fastapi>=0.124.4",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.1",
    "langchain-openai>=1.1.4",
    "langgraph>=1.0.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
